refactor: Change schema directory.
