+++ src/admin/fields/schemas/collectionFields/choice/createChoice.ts
@@ -1,0 +1,12 @@
+import { ObjectSchema } from 'yup';
+import yup from '../../../yup';
+
+export type FormValues = {
+  label: string;
+  value: string;
+};
+
+export const createChoice: ObjectSchema<FormValues> = yup.object().shape({
+  label: yup.string().required().max(60),
+  value: yup.string().required().max(60),
+});
